LayoutTests/http/wpt/web-animations/interfaces/Animation/startTime-expected.txt


PASS startTime of a newly created (idle) animation is unresolved 
FAIL startTime of a play-pending animation is unresolved assert_equals: startTime is unresolved expected (object) null but got (number) 61.900000000000006
PASS startTime of a pause-pending animation is unresolved 
FAIL startTime of a play-pending animation created using Element.animate shortcut is unresolved createDiv(t).animate is not a function. (In 'createDiv(t).animate(null)', 'createDiv(t).animate' is undefined)
FAIL startTime is resolved when running createDiv(t).animate is not a function. (In 'createDiv(t).animate(null, 100 * MS_PER_SEC)', 'createDiv(t).animate' is undefined)
FAIL startTime and currentTime are unresolved when animation is cancelled createDiv(t).animate is not a function. (In 'createDiv(t).animate(null, 100 * MS_PER_SEC)', 'createDiv(t).animate' is undefined)
